+++ CHANGELOG.md
@@ -1,3 +1,35 @@
+<a name="4.2.0"></a>
+# 4.2.0 (2017-04-24)
+
+
+### Bug Fixes
+
+* **calendar-input:** don't fire blur event when input loose focus and calendar will receive focus ([686eb32](https://github.com/alfa-laboratory/arui-feather/commit/686eb32))
+
+
+### Features
+
+* **popup:** add data-for attr ([#17](https://github.com/alfa-laboratory/arui-feather/issues/17)) ([ccd14a5](https://github.com/alfa-laboratory/arui-feather/commit/ccd14a5))
+
+
+
+<a name="4.1.0"></a>
+# 4.1.0 (2017-04-20)
+
+
+### Bug Fixes
+
+* **form-field:** radio/checkbox group in form-field case ([#4](https://github.com/alfa-laboratory/arui-feather/issues/4)) ([01ec77c](https://github.com/alfa-laboratory/arui-feather/commit/01ec77c))
+* **input:** fix input min-width with width='available' prop ([#8](https://github.com/alfa-laboratory/arui-feather/issues/8)) ([60bc94b](https://github.com/alfa-laboratory/arui-feather/commit/60bc94b))
+* **input-autocomplete:** default size ([5a38aed](https://github.com/alfa-laboratory/arui-feather/commit/5a38aed))
+
+
+### Features
+
+* **scroll-to:** set default duration to 0 ([#6](https://github.com/alfa-laboratory/arui-feather/issues/6)) ([7631de3](https://github.com/alfa-laboratory/arui-feather/commit/7631de3))
+
+
+
 <a name="4.0.1"></a>
 ## 4.0.1 (2017-04-14)